TranslatablePackageInterface // introduce interface to class. TranslatablePackageFactory // moved into "Package"-domain and merged ApiEndpointResolver and DirectoryResolver into it. tests/Unit // update unit tests.
